+++ src/main.jl
@@ -50,26 +50,44 @@ const vec_colormap = [[0, 0, 0],          # 0
 # [138, 138, 119]
 
 
-const voc_classnumbers = sort( Dict(zip(vec_classes, vec_classnames)) )   # harmonization with CocoTools
+const classnumbers = sort( Dict(zip(vec_classes, vec_classnames)) )   # harmonization with CocoTools
+function voc_classnumbers(class::Int)
+    return get(classnumbers, class, classnumbers[0])   # 0 is background
+end
 const voc_classnumber2classname = voc_classnumbers
 
-const voc_classnames = sort( Dict(zip(vec_classnames, vec_classes)); byvalue=true )   # harmonization with CocoTools
+
+const classnames = sort( Dict(zip(vec_classnames, vec_classes)); byvalue=true )   # harmonization with CocoTools
+function voc_classnames(name::String)
+    return get(classnames, name, classnames["background"])   # "background" is default
+end
 const voc_classname2classnumber = voc_classnames
 
-const voc_colormaps = sort( Dict(zip(vec_colormap, vec_classes)); byvalue=true )
-const voc_colormaps2classnumbers = voc_colormaps
-const voc_colormap2classnumber  = voc_colormaps
 
-const voc_classnumbers2colormaps = sort( Dict(zip(vec_classes, vec_colormap)) )
+const colormaps = sort( Dict(zip(vec_colormap, vec_classes)); byvalue=true )
+function voc_colormaps(colormap::AbstractVector{<:Integer})
+    return get(colormaps, colormap, colormaps[[0, 0, 0]])   # [0, 0, 0] is background
+end
+const voc_colormaps2classnumbers = voc_colormaps
+const voc_colormap2classnumber   = voc_colormaps
+
+
+const classnumbers2colormaps = sort( Dict(zip(vec_classes, vec_colormap)) )
+function voc_classnumbers2colormaps(class::Int)
+    return get(classnumbers2colormaps, class, classnumbers2colormaps[0])   # 0 is background
+end
 const voc_classnumber2colormap = voc_classnumbers2colormaps
+
 
 function voc_rgb2classes(mask::AbstractArray{RGB{N0f8}, 2})
     h, w = size(mask)
     X    = Matrix{Int}(undef, (h,w))
-    m    = channelview(mask) .* 255 .|> Int   # CHW
+
+    m1 = channelview(mask)   # CHW
+    m = permutedims(m1, (2,3,1)) .* 255 .|> Int   # HWC
 
     for i in 1:h; for j in 1:w
-        X[i,j] = get(voc_colormap2classnumber, ([m[1, i, j], m[2, i, j], m[3, i, j]]), 0)
+        X[i,j] = voc_colormaps2classnumbers([m[i,j,1], m[i,j,2], m[i,j,3]])
     end; end
 
     return X